Refactor: 당첨통계 결과 관련 클래스 분리
Result객체는 Result에 대한 값만을 관리하고
해당 결과를 계산을 담당하는 객체는 새로운 객체가 필요함
-> WinningStatic 객체 생성
